refactor: Correct inheritance
