+++ actfw_core/task/pipe.py
@@ -1,70 +1,25 @@
 import inspect
-import traceback
-from queue import Empty, Full, Queue
-from threading import Thread
-from typing import Generic, List, TypeVar
+from typing import Generic, TypeVar
 
 from ..util.pad import _PadBase, _PadBlocking, _PadIn, _PadOut
+from .consumer import _ConsumerMixin
+from .producer import _ProducerMixin
 from .task import Task
 
 T_OUT = TypeVar("T_OUT")
 T_IN = TypeVar("T_IN")
 
 
-class Pipe(Task, Generic[T_OUT, T_IN]):
-    running: bool
-    in_queues: List[_PadOut[T_IN]]
-    out_queues: List[_PadIn[T_OUT]]
-    out_queue_id: int
-
+class Pipe(Generic[T_OUT, T_IN], Task, _ProducerMixin[T_OUT], _ConsumerMixin[T_IN]):
     """Straightforward Pipeline Task."""
 
-    def __init__(self):
+    def __init__(self) -> None:
         """"""
-        super().__init__()
-        self.running = True
-        self.in_queues = []
-        self.out_queues = []
-        self.out_queue_id = 0
+        Task.__init__(self)
+        _ProducerMixin.__init__(self)
+        _ConsumerMixin.__init__(self)
 
-    def _is_running(self):
-        return self.running
-
-    def _add_in_queue(self, q):
-        self.in_queues.append(q)
-
-    def _add_out_queue(self, q):
-        self.out_queues.append(q)
-
-    def _inlet(self):
-        in_queue_id = 0
-        length = len(self.in_queues)
-        while self._is_running():
-            try:
-                i = self.in_queues[in_queue_id].get(timeout=1)
-                yield i
-                in_queue_id = (in_queue_id + 1) % length
-            except Empty:
-                pass
-            except GeneratorExit:
-                break
-            except:
-                traceback.print_exc()
-
-    def _outlet(self, o):
-        length = len(self.out_queues)
-        while self._is_running():
-            try:
-                self.out_queues[self.out_queue_id].put(o, timeout=1)
-                self.out_queue_id = (self.out_queue_id + 1) % length
-                return True
-            except Full:
-                pass
-            except:
-                traceback.print_exc()
-        return False
-
-    def run(self):
+    def run(self) -> None:
         """Run and start the activity"""
         for i in self._inlet():
             o = self.proc(i)
@@ -72,29 +27,8 @@ class Pipe(Task, Generic[T_OUT, T_IN]):
             if not self._is_running():
                 break
 
-    def proc(self, i):
+    def proc(self, i: T_IN) -> T_OUT:
         """
         Pipeline Task Processor
-
-        Args:
-            i : task input
         """
         raise NotImplementedError("'proc' must be overridden.")
-
-    def stop(self):
-        """Stop the activity"""
-        self.running = False
-
-    def connect(self, follow):
-        """
-        Connect following task.
-
-        Args:
-            follow (:class:`~actfw_core.task.Task`): following task
-        """
-        pad_out, pad_in = self._new_pad().into_pad_pair()
-        follow._add_in_queue(pad_out)
-        self._add_out_queue(pad_in)
-
-    def _new_pad(self) -> _PadBase[T_OUT]:
-        return _PadBlocking()
